Move read_password from main.c to userinput.c
Preparation for reuse in the 2factor implementation.
